update for 1.20.1 api

+++ src/main/java/io/yukkuric/hex_ars_link/items/ItemLinker.java
@@ -1,8 +1,9 @@
 package io.yukkuric.hex_ars_link.items;
 
 import at.petrak.hexcasting.common.items.magic.ItemMediaHolder;
+import com.hollingsworth.arsnouveau.api.mana.IManaCap;
 import com.hollingsworth.arsnouveau.api.util.ManaUtil;
-import com.hollingsworth.arsnouveau.common.capability.CapabilityRegistry;
+import com.hollingsworth.arsnouveau.setup.registry.CapabilityRegistry;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.InteractionHand;
@@ -42,21 +43,21 @@ public class ItemLinker extends ItemMediaHolder implements OwnerBinder {
     }
 
     @Override
-    public synchronized int getMedia(ItemStack stack) {
+    public synchronized long getMedia(ItemStack stack) {
         var player = getOwner(stack);
         if (player == null) return 0;
-        return (int) (ManaUtil.getCurrentMana(player) * getConvertRatio());
+        return (long) (ManaUtil.getCurrentMana(player) * getConvertRatio());
     }
 
     @Override
-    public synchronized int getMaxMedia(ItemStack stack) {
+    public synchronized long getMaxMedia(ItemStack stack) {
         var player = getOwner(stack);
         if (player == null) return 0;
-        return (int) (ManaUtil.getMaxMana(player) * getConvertRatio());
+        return (long) (ManaUtil.getMaxMana(player) * getConvertRatio());
     }
 
     @Override
-    public synchronized void setMedia(ItemStack stack, int i) {
+    public synchronized void setMedia(ItemStack stack, long i) {
         var player = getOwner(stack);
         // forge only
         CapabilityRegistry.getMana(player).ifPresent(mana -> mana.setMana(((double) i) / getConvertRatio()));
